requirements.txt
discord.py
python-dotenv
uvloop